[dv] enable copying filegroup targest to run dir
When a filegroup is used to point to a SW image in a dvsim config file,
and the filegroup lives in an external bazel repo, i.e.,
@manufacturer_test_hooks//..., excuting a `bazel cquery ...` operation
on it to obtain the file paths will yield a partial path relative to the
`bazel info output_base` path. We need to account for this in the sim.mk
file when copying over Bazel build SW images to the dvsim run dir.

Signed-off-by: Tim Trippel <ttrippel@google.com>