[reggen] Re-order subfields of hw2reg packed structs
Reverse the order of the sub-fields within hw2reg packed
structs to mirror the change that was made to reg2hw in
61a237e.

This helps with consistency but also addresses a direct
structure assignment in soc_dbg_ctrl.

Update the assignment to the status register within
otp_ctrl accordingly because it assumes the field ordering.

Signed-off-by: Adrian Lees <a.lees@lowrisc.org>