[racl_ctrl,rtl] Set error_log's valid and overflow bits correctly
I think the idea of the design was that the hardware might report a
racl error, which causes the valid bit to latch high until the
software writes to the valid field of the error_log register.

The existing code didn't quite do that. The logic seems to have been
that we should clear the log on the second write of an error (which
includes writes from reported errors from other blocks). As a result,
a misconfigured DV sequence that reports an error to racl_ctrl for
several cycles will see an oscillating valid signal. That didn't seem
right to me!

With the new version of the RTL, SW is only able to clear errors and
other blocks are only able to report them.

When HW blocks report an error, it gets arbitrated and appears in the
racl_error_arb signal. If ERROR_LOG.VALID isn't already true, the
first_error signal gets set, which is used to control the .d value
that will be written into the registers.

SW clears an error by writing 1 to the ERROR_LOG.VALID field. That
field is rw1c so the auto-generated register implementation handles
that. Other fields will get cleared on the next cycle (because
the had_error signal will be false).

Notice that this means software theoretically read ERROR_LOG register
with e.g. ERROR_LOG.OVERFLOW && !ERROR_LOG.VALID. To do this, it would
need to write to the error_log register and read from it again on two
consecutive cycles.

But it is still not possible to "latch a spurious value": an error
being reported on that last cycle wouldn't stop the data field from
being cleared, because *that* only depends on the valid flag having
been true.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>