[racl_ctrl,dv] Report occasional error logs in racl_ctrl_base_vseq
This is done using the internal and external error log agents. The
body of the base vseq spawns sequences on those two agents to report
the occasional error.

We also have a background thread that checks occasionally and does a
TL write to clear the error log if it is valid.

When the sequence body completes (which probably takes a significant
time in the child class), the post_start task runs. This actually
blocks, asking both child sequences to stop and stops sending error
log clear requests.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>