[opentitanlib] add option to ignore SPI console frame numbers
When opentitanlib is invoked via an external C/C++ wrapper library, the
SpiConsoleDevice object is reinstantiated repeatedly to comply with
Rust's ownership rules (and satisfy the borrow checker). On each
reinstantiation of the SpiConsoleDevice struct, the next expected frame
number is reset to 0. This causes the SpiConsoleDevice to ignore a valid
frame when used through an external wrapper library, as is the case with
the OtlibWrapper used in the opentitan-provisioning repo. This was
causing CI issues in
https://github.com/lowRISC/opentitan-provisioning/pull/165.

As such, this provides a mechanism for users of the SpiConsoleDevice to
ignore the frame number, which frankly  is not necessary when the
SpiConsoleDevice is used alongside the TX-indicator GPIO feature.
This feature does not make use of a circular transmit buffer, but rather
transmits only a single frame in the buffer at a time.

Signed-off-by: Tim Trippel <ttrippel@google.com>
(cherry picked from commit 9ca5d547d270373cc86a12583ed3f5148f3cf682)