[util] Add filter for existing committers to contributor stats
Add a command-line option to filter out those who are already in the
COMMITTERS file from the generated list of contributor stats, so that it
can be used to identify potential candidates to become committers.

This is limited however in that it only looks for current contributors -
anybody who was previously a committer but is not any more will still
be suggested for now.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>