[util] Find historical committers for committer stats
Extends existing contributor stats script logic to find and filter
committers in the COMMITTERS file from the results, to now find a
complete list of all committers in the (tracked) history of the repo, by
searching the local git revisions of changes and parsing the file
contents at each point.

This is more useful, as it means that using the `--filter` option will
no longer return several historic committers.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>