[prim,rtl] Add a NeverClears parameter to prim_fifo_sync(_cnt)
This can be used by instantiations to assert that the clr_i port is
never true (passing that information to the prim_fifo_sync_cnt and
then the prim_count inside).

The commit is quite large, but that's because there are lots of
instantiation sites. If one of these parameters is passed wrongly, the
result will be caught immediately at simulation (from the
NeverClears_A assertion that's added to prim_fifo_sync or the
ClrNeverTrue_A assertion that this turns on inside prim_count).

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>