]provisioning] Allow to mix in SKU specific success text
When wating for successful post personalization restart in some cases
the text printed on the device console is not static and depends on
data generted during personalization.

Modify ft_ext() prototype to accept the response structure and to
return a String Option which the host would mix in with the constant
success string when defining the expected console output.

Tested with an updated Cros image which generates dev seed report.

Signed-off-by: Vadim Bendebury <vbendeb@google.com>