[ci,blockfile] Enforce two distinct commiters
For authorising changes, one committer could send in two approvals. This
commit enforces that the authorisation must be from unique committers.

Sets in Python enforce unique values, while lists allow duplicates. This commit changes the return type of `get_authorized_changes` to be a set instead of a list. Later the length of this set is taken which guarantees that each committer is only counted once.

Co-authored-by: Rupert Swarbrick <rswarbrick@lowrisc.org>
Signed-off-by: Marno van der Maas <mvdmaas+git@lowrisc.org>