[topgen] Add support for dtgen in ip config
In order to export ipgen parameter to dttool, a new 'dtgen' attribute
is added to the tpldesc file of ipgen IPs. Its content is simply
forwarded as-is to the generated ipconfig file. For example:
```hjson
{
      name: "num_inp_period_counters"
      desc: "number of input period counters"
      type: "int"
      default: 0
      dtgen:
      {
        type: "uint8"
        name: "input_period_counter_count"
      }
}
```
will generate
```hjson
{
  instance_name: top_earlgrey_gpio
  param_values:
  {
    num_inp_period_counters: 0
    ...
  }
  dtgen:
  {
    num_inp_period_counters:
    {
      type: uint8
      name: input_period_counter_count
      // This attributes gets added by topgen
      doc: number of input period counters
    }
  }
}
```
The only nontrivial thing that topgen does is copy the `desc` field
of each parameter into the `doc` attribute of dtgen. The content of
the `dtgen` attribute is not validated because it is the responsability
of `dttool` to interpret its content

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>