[dttool] Add support for multiple extensions
The code was assuming the existence of a unique extension (which
was set to fake empty extension if none is provided). This commit
adds support for multiple extension. The main change required is
that each extension must now indicate the name of the field in
the structure that should be used to store its data.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>