[dtgen] Add ipgen extension to expose ipconfig parameters
This extension will look at the ipconfig parameters and automatically
generate code for those parameters marked with `dtgen`. The format
of the `dtgen` dictionary is currently as follows:
- `type`: type of the data exposed by the dt API. Only `uint8` is
  currently supported but the code is architectured in a way that
  it can easily support any scalar type
- `name` (optional): name to use (otherwise the parameter name)
  This controls the API which will be named `dt_<ip>_<name>()`
- `doc`: optional documentation of the parameter.

The most important part of the generated code is a new dt function
for each attribute of the form:
```c
<type> dt_<ip>_<name>(dt_<ip>_t dt);
```
which returns the parameter value (or an set invalid value if the dt
argument is invalid). Here is an example for the GPIO:
```c
/**
 * Get the number of input period counters.
 *
 * @param dt Instance of gpio.
 * @return number of input period counters.
 */
uint8_t dt_gpio_input_period_counter_count(dt_gpio_t dt);
```

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>