[ottf] Introduce a new OTTF console state structure
Currently, the console code casts UART and SPI DIFs to void* pointers
and passes them around to the various sink, putbuf and getc functions.
This makes the code difficult to maintain and prevents pasing extra
information about the console.

This commit introduces a new `ottf_console_t` struct which replaces
the DIFs in the void* pointers mentioned above. This structure currently
contains the console type and the DIF structure. This requires a few fixes
here and there, and the various configuration functions of console needs
extra pointers to the console object.

To replace the old DIF struct which were previously global variables,
a global "main console" object is created in `ottf_console.c`.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>