[rules,otp] Create new rules that allows collect OTP overlays
Currently overlays must all be specified as a list when calling
`otp_image`. This is inconvenient because it means that those
lists have to be created in .bzl files instead of BUILD files.

This commit introduces a new `otp_json_overlay` rule whose purpose
is to simply collect overlays so that they can be refer to as a
single target. The `otp_image` rule is modified to accept those,
along `otp_json` files.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>