[dv, alert_handler] Remove `illegal_bins` where not needed
Alert handler defines specific bins for the coverpoints. So when X-coverage
takes place, only the explicitly defined bins will be crossed, and hence the
illegal_bins won't influence the cross.

A check has been moved the the SCB to be active regardless of whether coverage
collection is enabled or not. These checks just ensure the TB is configured
within the design bounds by making sure the analysis imports for alert and esc
are within the design bounds.

Signed-off-by: Antonio Martinez Zambrana <amz@lowrisc.org>