[sram_ctrl] Increase width of LFSR used for initialization to 64 bits
Previously, this LFSR just used a width of 32 which - depending on how
the module is used - may not be ideal. This commit thus increases the
width of this LFSR to 64 bits, while the output width, i.e., the number
of bits used per cycle, remains at 32 bits. To seed the additional 32
state bits, the previously unused upper 32 bits of the nonce obtained
from OTP_CTRL are used.

This is related to lowRISC/OpenTitan#27381.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>