[perso] log perso firmware has in the perso blob
Prior to this change, the personalization firmware (self) hash was
logged to the console as text. This updates the flow to push it into the
perso TLV structure (i.e., perso blob) so it can be sent to the
provisioning registry.

Signed-off-by: Tim Trippel <ttrippel@google.com>
(cherry picked from commit 94fdaf220d0ffdabcb9d2d778e74fba782f42174)