[sival, kdf] Compare derived keys in kdf_kmac_sideload_functest
The `kdf_kmac_sideload_functest` computed two blinded keys using KDF
with sideloaded keys from the keymgr, but directly compared the output
blinded keyblob, which will contain two masked keyshares.

KMAC masking is disabled on CW310 due to FPGA limitations, but in other
execution environments we would not expect the two keyblobs to be the
same as a result of KMAC masking.

To fix this, correct the test logic to export unmask both derived
blinded keys, and directly compare the contents of the unmasked keys
instead. This should allow the test to pass in environments that have
KMAC masking enabled.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>