[tests,usb] More spec compliant suspend/resume
Per the USB hub specification, when performing port operations, we
should look at the port status to make sure that the operation was
performed. This replaces the artificial 100ms delay which does not
correspond to any real specification timing.

This commit also introduces stricter checks on the port status:
an error will occur if trying to suspend/resume a port with
incompatible status (though even stricter checks are possible such
as making sure that a device is connected).

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>