[tests,usb] Do not suspend in the Shutdown phase
When the device enters the shutdown phase, it will very quickly
configure the usbdev interface, try to connect and very shortly
afterwards disable the interface. This means that the device may
or may not appear on the bus (most likely not). Meanwhile, the
host is trying to suspend the device in this phase which is incorrect.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>
(cherry picked from commit c0863cdd79497ae82f792f590e0444d264a4a0a1)