[tests,usb] Make the VBUS sense delay configurable
When enabling/disable VBUS sense, it is necessary to wait for a
while for things to stabilize. Make this delay configurable and
bump it to 200ms since 100ms is too low on certain boards.
It is not clear why but there is no real harm in increasing this
delay.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>
(cherry picked from commit b8a0597921d553f4bcc323c1a4c5154fa089d406)