[prim,fpv] Add a "helper assertion" for prim_diff_decode
When doing FPV, the model checker can end up searching in an
unreachable state where it happens to have started with a differential
decoder in state 2'b11. To avoid it wasting lots of time looking at
"stuck" unreachable states, prove that this can't happen.

The "after_load" script registers this property as a helper, which
should mean state space exploration for things driven by the level of
this signal doesn't get stuck looking at unreachable states where
state_q is a constant 2'b11.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>