[fpv] Make sequence chains compile much faster
If you have a sequence that looks like "A ##1 B ##1 C ##1 D", Jasper
turns out to take quite a long time to compile it. This is rather
annoying, and you can actually hit a timeout.

It seems that the Cadence engineers realised that they could do things
more efficiently and put it behind a flag (presumably because the
change pessimizes some other example).

Ask for the new behaviour, provided we are on a version of Jasper that
supports it.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>