[crypto,sw] Add is on curve checks to scalar_mult_int
This commit adds is on curve checks each time scalar_mult_int
is executed. This serves as an FI countermeasure. In case a fault
is injected to get an invalid point such that the secret can be
leaked this will be caught by the is on curve check.

Signed-off-by: Hakim Filali <hfilali@lowrisc.org>
(cherry picked from commit 22d99a4f3d6801b64a1356ab5fcc7d1bf11af11d)