[alert,fpv] Simplify AlertCheck1_A in rxtx async assert
The original version of this assertion said "if I ask the alert sender
for an alert then eventually the alert receiver will report it".
Commit c8fb0e6 made this rather more complicated by requiring the
sender to go back to the idle state before that happens.

This may well be true, but I can't really see any need (and it rather
mixes the two ends' timelines). Simplify the assertion again.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>