[rv_plic, fpv] Tweak FpvSecCmRegWeOnehotCheck_A assertion
This assertion says that if alert happens then within a bounded
time, the sender will send alert_p. To make this true, we
required alert receiver and a whole world of additional FPV for
alert handling in rv_plic. Also, it will make more sense to
expect an alert request coming from rv_plic to the sender
rather then expecting the sender to send the response.

Signed-off-by: Kinza Qamar <kqzaman@lowrisc.org>