[alert,fpv] Make AlertCheck1_A actually true for async alerts
This assertion wasn't true because a skew on e.g. the ack signal from
the alert receiver to the alert sender can completely lock up
communication between them.

It seems that Jasper didn't manage to find the loop it would need for
a counterexample with this sort of liveness proof, so it just hung.
With the extra condition (which, in hindsight, is definitely
required), the proof goes through nicely.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>