[sw/tests] kmac_smoketest unaligned sha3 check fix
There were two separate issues here, firstly the message length for the
unaligned SHA3 test was including the null character which this commit
fixes by declaring the message as a character array instead of a string.

The second issue was that the output was being checked using
CHECK_DIF_OK instead of CHECK. These have inverted behavior which meant
the mismatch was never reported.

Signed-off-by: Marno van der Maas <mvdmaas+git@lowrisc.org>