[alert,fpv] Strengthen AlertHs_A in prim_alert_rxtx_assert_fpv
Before this commit, the property in the AlertHs_A assertion looked
something like

  alert_req_i && both_idle |=> FullHandshake_S |-> alert_ack_o

Noting that |-> and |=> are right-associative, this means

  "Suppose there's an alert request when both are idle. Then any full
   handshake that starts on the following cycle will end with
   alert_ack_o being high"

That seems rather weak! I think the author intended it to say that
there would be a full handshake and that handshake would end with
alert_ack_o being high.

With the rewording in the commit, this becomes:

  "Suppose there's an alert request when both are idle. Then a full
  handshake will start on the following cycle, finishing with with
  alert_ack_o being high"

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>