[sw] Add `nonstring` attribute
Adds the `nonstring` attribute through a new `OT_NONSTRING` macro.
This is used to prevent warnings when compiling with Clang 21+ and an
array is initialized with a string literal but is sized not to
include the terminating null.

Signed-off-by: Luís Marques <luismarques@lowrisc.org>
(cherry picked from commit c64f1fa935e93a68057b511cc65507af009e5a39)