[sw] Add `nonstring` attribute
Adds the `nonstring` attribute through a new `OT_NONSTRING` macro.
This is used to prevent warnings when compiling with Clang 21+ and an
array is initialized with a string literal but is sized not to
include the terminating null.

Signed-off-by: Luís Marques <luismarques@lowrisc.org>
(cherry picked from commit 2df6501c63beb1f2bd11e5e5d315e3a903fbebd3)