[util] Teach check_tool_requirements to work properly with fusesoc
fusesoc doesn't actually register its version with importlib in the
way we expected, so it comes out with version '0.0.0' if you run with
a Python that can see the head of the upstream branch. Switch the code
in check_tool_requirements to ask in a way that it can get a sensible
answer.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>