[uart,dv] Do not error on missing UARTDPI_LOG_* plusarg
Current Verilator simulation does not provide a plusarg for the UARTDPI
logs and relies on the default log files being used (e.g. "uart0.log").
The introduction of an error on a missing plusarg causes Verilator
simulation builds to fail - for now, we just `display` this error
instead as a useful piece of info.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>