[rom_ext] set DIN explicitly in the rom_ext_boot_services_unittest
The C++ standard only guarantees the initialization of the first member of a union, which in this case only covers the header of the boot service message. This behavior, when compiled with Clang, leads to test failures.

This change explicitly sets DIN to ensure the value matches the lifecycle device ID.

Change-Id: I60e4b77a0733ef23016d50b69af26eb50c239c67
Signed-off-by: Yi-Hsuan Deng <yhdeng@google.com>