[tests,usb] Drain UART on failure in usbdev_suspend
When the host fails due to some error (like hub status failure),
the UART output will not be printed. This can make debugging the
test challenging since some crucial information about the device
state is lost. Change the test so that in case of failure, the
UART is drained and displayed.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>