[tests,usb] Increase usb_suspend resume timeout
Currently there is a sequence where the devices prints a message,
suspends and then waits for a resume. On the host side, we need
to wait for the UART message, send a request to the hub to suspend,
observe that the suspend was successfull and then resume. The current
timeout between suspend and resume is only 40ms which can be quite
tight since hub operations can take a long time. Increase this
delay to avoid flakiness.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>