[sw, opentitanlib] Warn on Hyperdebug UART baud rate mismatch
When setting the baud rate on Hyperdebug via the USB stream, we divide
by 100, and the Hyperdebug firmware then performs the corresponding
multiplication to retrieve the original baud rate. Perform the
calculation on the host as well and check that the decompressed value
matches the initial value, emitting a warning if some data has been
lost.

Probably in practice no baud rate that can be used with Hyperdebug will
face this issue, but it's still better to encode for this edge case than
rely on it implicitly working.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>
(cherry picked from commit ca617461bab348a661c0d3efdb879fdff1e3cf54)